+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/chat/model/UnReadBo.java
@@ -3,7 +3,7 @@ package cc.xmist.mistchat.server.chat.model;
 import cc.xmist.mistchat.server.common.enums.RoomType;
 
 public class UnReadBo {
-    private Long uid;
+    private Integer uid;
     private RoomType chatType;
-    private Long chatId;
+    private Integer chatId;
 }
