+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/common/config/resolver/UserIdArgumentResolver.java
@@ -19,7 +19,7 @@ public class UserIdArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterName().equals("uid") && parameter.getParameterType().equals(Long.class);
+        return parameter.getParameterName().equals("uid") && parameter.getParameterType().equals(Integer.class);
     }
 
     @Override
