+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/common/interceptor/BlackInterceptor.java
@@ -19,9 +19,9 @@ public class BlackInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        Long uid = RequestContext.getUid();
+        Integer uid = RequestContext.getUid();
         List<String> blockedIp = blackService.getBlockedIp();
-        List<Long> blockedUid = blackService.getBlockedUid();
+        List<Integer> blockedUid = blackService.getBlockedUid();
 
         String ip = JakartaServletUtil.getClientIP(request);
