+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/group/req/GroupCreateRequest.java
@@ -15,5 +15,5 @@ public class GroupCreateRequest {
 
     @Length(min = 0)
     @Schema(description = "用户列表")
-    private List<Long> uidList;
+    private List<Integer> uidList;
 }
