+++ mistchat-server/src/main/java/cc/xmist/mistchat/server/user/req/UidListRequest.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class UidListRequest {
     @Schema(name = "用户 id 列表")
-    private List<Long> uidList;
+    private List<Integer> uidList;
 }
